requirements.txt
pillow
django
python-dotenv
flask